requirements.txt
requests>=2.25.1
twilio>=6.45.4
transformers>=4.0.0
redis>=3.5.3
pyyaml>=5.4.1
SpeechRecognition>=3.8.1